manager/pyproject.toml
[project]
name = "fw-manager"
version = "0.0.1"
description = "An album manager."
authors = [
    {name = "CodeKitchen", email = "allen@tkzt.cn"},
]
dependencies = [
    "bootstrap-flask>=2.3.3",
    "python-dotenv>=1.0.0",
    "flask-wtf>=1.2.1",
    "flask-sqlalchemy>=3.1.1",
    "click>=8.1.7",
    "loguru>=0.7.2",
    "pillow>=10.2.0",
    "blurhash-python>=1.2.1",
    "flask-httpauth>=4.8.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
package-type = "application"

[tool.pdm.scripts]
serve = { cmd = "flask run", help = "Run the development server" }
test = { cmd = "pytest", help = "Run all test cases" }
create-tables = { cmd = "flask create-tables", help = "Create tables" }
drop-tables = { cmd = "flask drop-tables", help = "Drop tables" }

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.4.4",
    "pytest-cov>=4.1.0",
]
